refactor: formatTime(날짜 포맷 변환 함수) shared/utils로 분리
